refactored and moved state eliminated a component that wasn't really doing anything
