+++ project-3-front/src/PageContainer/ListComponent/ListItem.js
@@ -6,27 +6,24 @@ import React from 'react'
 //Starting with "../../" moves two directories backwards and starts there(and so on...)
 //To move forward, just start with the first subdirectory and keep moving forward
 
-const ListItem = (props) =>{
-    const campsite = props.campsites.map((campsite) => {
+
+const ListItems = (props) => {
+
+    const campsiteList = props.campsites.map((campsite) => {
         return (
             <li key={campsite._id}>
-                <p>{campsite.name}</p>
+                <div>{campsite.name}</div>
             </li>
-        );
-    });
+
+        )
+    })
 
     return (
         <ul>
-            {campsite}
+            {campsiteList}
         </ul>
-    );
-    
+    )
 
 }
 
-
-
-
-
-
-export default ListItem
+export default ListItems;
