+++ project-3-front/src/PageContainer/MapContainer/MyMapComponent.js
@@ -20,38 +20,8 @@ const MyMapComponent = compose(
         defaultZoom={8}
         defaultCenter={{ lat: 39.7392358, lng: -104.990251 }}
     >
-        {props.isMarkerShown && <Marker position={{ lat: -34.397, lng: 150.644 }} onClick={props.onMarkerClick} />}
+        {props.isMarkerShown && <Marker position={{ lat: 39.7392358, lng: -104.990251 }} onClick={props.onMarkerClick} />}
     </GoogleMap>
 )
-
-class MyFancyComponent extends React.PureComponent {
-    state = {
-        isMarkerShown: false,
-    }
-
-    componentDidMount() {
-        this.delayedShowMarker()
-    }
-
-    delayedShowMarker = () => {
-        setTimeout(() => {
-            this.setState({ isMarkerShown: true })
-        }, 3000)
-    }
-
-    handleMarkerClick = () => {
-        this.setState({ isMarkerShown: false })
-        this.delayedShowMarker()
-    }
-
-    render() {
-        return (
-            <MyMapComponent
-                isMarkerShown={this.state.isMarkerShown}
-                onMarkerClick={this.handleMarkerClick}
-            />
-        )
-    }
-}
 
 export default MyMapComponent
